convolve: use pointer arithmentics
Avoding indexed acces on the source, target and kernel arrays should yield
better performance when the kernel and/or the source image are big enough. I
tried the current version of the code in my Atom(TM) CPU N550 and found a speed
up from 4.8 [s] to 4.0 [s] when convolving a 1000x1000 image with a 16x16
kernel.

The code still has to be thoroughly re-read to make sure it's doing the right
thing. I'm also a bit unsure whether I could come up with a more interesting
iteration over the source image.

Signed-off-by: Rodrigo Javier Tobar Carrizo <rtobarc@gmail.com>