sersic: fix minor-axis gradient calc, adjust acc
Calculations of the gradient in the minor axis didn't consider the axis ratio
and therefore were not as good as they should have been. Changing it as it how
it should be means we also have to adjust the accuracy to a proper value
depending on the sersic index and the axis ratio; otherwise we would spend too
much time on sub-pixel integration.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>