convolve: kernel wasn't flipped
The kernel has to be flipped in both directions when convolving; the previous
version of the code iterated over the kernel the same way as the image, and
therefore the result was the convolution of the image with the actual flipped
kernel. I've changed the iteration order of the kernel so now the result is
really the convolution of the image with the kernel.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>