Transitioning to cmake
This means we no longer use the simplistic hand-written src/Makefile, but need
to run cmake first and then generate the build system.

The new cmake approach automatically scans for GSL and R, and uses one or the
other in the correct order. The docs have been updated to show how to compile
libprofit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>