psf: new code handling resolutions
The previous code was not handling image and psf resolution at all, assuming
that both the psf and the image had the same vertical and horizontal
resolution, and that they were equals to 1. This new code (still to be fully
tested) takes all these variables into account, making it more complex but more
accurate.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>