serisc,moffat,ferrer: merged common code
A new profile class named SersicLikeProfile has been created containing the
code that is common to all three Sersic, Moffat and Ferrer profiles. The code
abstracts away all the commonalities, leaving the specifics on each subclass.

This first iteration compiles cleanly and seems to work as well as before but
probably better testing is required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>