sesic-like: re -> rscale
The effective radius (re) is not a common concept across all the sersic-like
profiles, but instead a sersic-specific feature (for the moment). We have thus
renamed this common "re" variables (re, re_max and re_switch) to "rscale"
(rscale, rscale_max and rscale_switch) to avoid confusions in the future.

This also means that now "re" lives only in the sersic profile.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>