analytic: renamed from sersic_like
The old SersicLikeProfile class has been renamed to AnalyticProfile to avoid
name clashes and confusions. The name might change later on again as it
probably doesn't describe well enough what the commonality between all the
subclasses is, but on the other hand we don't want to end up with a class named
EllipticalAndSubpixelIntegratedProfile.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>