utils: check domain of gsl-based gamma function
From the GSL manual:

"The maximum value of x such that gamma(x) is not considered an overflow is
given by the macro GSL_SF_GAMMA_XMAX"

So instead of blindly passing down all values we first check whether it will
make an appropriate response or not.

The R-based gammafn does this already automatically so we don't need to worry
on that side.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>