GSL funcs behave like their R counterpart
The beta and gamma functions work slightly differently to their R counterparts.
To emulate this behavior in the GSL functions we need to carefully look into
the parameters being passed down, and into the status returned by the function
evaluation, and return the same values R does on each case.

To make sure this is working fine I've added some unit tests.

Also importantly, for GSL to work without aborting the program upon an error
(its default behavior) users need to set the gsl error handler off (or set a
custom handler). This is being done by the new unit tests, but we still need to
think of a clever and abstract way to expose this to the users, if at all.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>