Better encapsulation in profile classes
This commit redesigns the evaluation and integration functions used by the
different radial profiles so they are not static functions defined on each
module file, but instead they are member functions of their classes. This way
they can access the class members directly instead of having to receive an
external reference to the profile object and having to inspect it.

This is the first step towards making the different profile parameters
protected. With this in place, a generic parameter setting interface will be
provided by the Profile class, and therefore users will not need to include
specific header files to use specific classes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>