Generic way to set parameters on profiles
This new set of methods on the Profile class allow users to set parameters in
profiles in a generic way, removing the need for static_cast<> the result of
model.add_profile and #include specific header files.

The tests have been updated to reflect this change as well.

This was also the final step towards proper encapsulation. Now that parameters
are not accessed externally by users we no longer need to keep them public.
They have been moved to protected instead, so classes can access the parameters
from their parent class.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>