Test-specific instructions to tests CMakeLists
I'm also preparing the tests for reusing code. This needs adding the --have-eh
argument for the cxxtestgen command to always define the CXXTEST_HAVE_EH macro,
and not only when the related TS_* assertions are found in the test code (which
doesn't happen when the TS_* assertions are used in a common .h file which is
later included in the main test class .h file).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>