Renaming header include guard macros
This is simply to be pedantic about the fact that macros starting with an
underscore and followed by either another underscore or an uppercase letter are
reserved identifiers for the standard C library. The new names are nicer
anyway, as they reflect better the namespace in which they are declared.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>