Removing the r and reuse_r radial eval params.
These were originally used to avoid some calculations on the sersic profile,
but when used in optimized builds the extra "if"s were actually in the way of
being efficient so I'm removing them. An option (I guess) would be to have the
two options with a compile-time switch to define which strategy to use.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>