Removing usage of lambdas
Because these functions were called from within the innermost loop of the
radial profile they were actually imposing a bit of an overhead on the
runtimes. I've changed the radial classes so their evaluate_at method is called
directly (which is the nicest thing to do, of course). The sersic class (the
source of all this added complexity because of its templated special cases)
instantiates internally the templated functions with all the different
parameters when overloading the main evaluate method. This function pointer is
stored in the class and used to call the corresponding templated function at
evaluate_at() time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>