cmake < 3.4 fails if only C++ is declared
Deep within cmake the check_symbol_exists macro fails to work correctly if C is
not specified as one of the languages of the project. This was fixed in 3.4, so
for cmake versions lower than that we have to include C as one of the languages
of the project to avoid the error.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>