Improving memory management
The get_opencl_environment function was previously returning a bare pointer to
the user, which then had to be manually deleted. This is very error prone, and
thus I've changed the function signature to return a shared pointer. The
reference is shared because the environment can be used by many
Models/Profiles. Thus, the reference stored by the Model has also become a
shared_ptr.

Following this trend I also changed the add_profile method of the Model class
to return a shared_ptr to the new Profile instead of a reference.

With these changes the Model class doesn't need an explicit destructor defined.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>