More OpenCL information available
The supported OpenCL version of the platforms is now queried at runtime, either
for display or for later usage. The double precision support of each device is
also now shown when querying for OpenCL system information.

Due to these changes, the way the OpenCL information is shown by the profit-cli
has also been enhanced.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>