Changed how radial profiles expose OpenCL support
Radial profiles now return a boolean value indicating whether they support
OpenCL or not. If they do, it is assumed that there will be a kernel called
<profile>_<float_type>, with float_type being either "float" or "double".

The patch is not properly isolated from other changes, but many changes had
been accummulated locally already and it was hard to separate things cleanly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>