Subsampling producing sensible results
Now the resulting image values are being stored and at the end appended to the
final image. For this to work a new structure (and a vector with them) was
introduced to store the temporary results, which are then processed and added
to the corresponding pixel of the final image.

Also in this commit there is a small try/catch to avoid memory issues when
trying to allocate a buffer on the target OpenCL device that is too big. In
that case we simply break from the recursion and live with the current result.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>