Subsampling, second version
There are two big changes in place here: first, both the data structures and
the variables holding subsampling data have been renamed for easier reading. It
should be clearer to the reader now what the code is doing.

Secondly, this second subsampling version uses less memory in the target device
-- not all the subsampling information is copied over to the target device
because it's actually not needed. The only bits that are needed (that is, the
x/y coordinates and the xbin/ybin sizes) are extracted out, copied over to the
target device, and then copied back to the host for processing.

The resulting image is actually wrong, but fixing it will go in a different
commit.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>