Upgrading to newer cl2.hpp
This version contains a fix on the CommandQueue constructors that manifested
itself with CL_HPP_TARGET_OPENCL_VERSION >= 200 and
CL_HPP_MINIMUM_OPENCL_VERSION < 200. The actual fixed version of the file is
still only available under github.com:rtobar/OpenCL-CLHPP, but a pull request
has been made to merge those changes back to the main repo, and hopefully get a
new release of the headers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>