libprofit 1.2
This new version is the first one to expose OpenCL capabilities. OpenCL >= 1.1
is required for things to run (although we might even do with OpenCL >= 1.0,
haven't tested that yet).

Other than that, there's better statistics reporting from the profiles up to
the caller. Things are not perfect, but could be worse.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>