Fixing sersic evaluation in some boxy cases
Our template specializations for the box != 0 case were being wrongly applied.
They were supposed to be specialized for X in the "(r/Re)^(1/X)" formula, where
X is nser*(box+2), but we were specializing them for nser instead, and thus in
the very few combinations of box != 0 && nser in [0.5, 1, 2, 3, 4, 8, 16] we
were giving an incorrect result. Note that for any other nser results were
correct. Also, the non-boxy case was also correct, because X in the formula
above is equals to nser.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>