Addig OpenCL-related tests
These tests are fairly simple, simply checking that an image produced for a
given profile using CPU-only evaluation doesn't vary much from that produced
using an OpenCL kernel. We use a default 5% relative difference tolerance to
declare success or failure, although this can be controlled by an environment
variable to be more or less strict.

The profile currently presenting more differences is coresersic, although still
within the 3% mark (for the given parameters used in the tests, which were all
extracted from the original profit paper).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>