Use R/RMath library dirs to look for libraries
These libraries can be in non-standard locations, and therefore need to
be added to the list of directories to search for libraries in order to
successfully link against them.

This command is rarely necessary because locations returned by
find_package and find_library usually are absolute paths. In our case
however we use pkgconfig to look for R and RMath, and therefore need to
take care of this ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>