First version of new config.h file
The new config.h file (created from config.h.in) records the compile-time
options used for libprofit. This way people using the library can check what
the library they are compiling against has support for, at compile-time.

Some of the macros have also been renamed to have a more uniform naming scheme
for all of them.

Finally, the new config.h file needs to be included in our code before checking
any of the macros (as opposed to simply assume they are always there, which was
the case when they were still given on the command-line). This includes .h
files, .cpp files and files under the test/ directory.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>