New Convolver classes
There are two classes implementing the Convolver interface. The
BruteForceConvolver implements the already-existing brute-force convolution
(and thus the profit::convolve function now simply creates a
BruteForceConvolver and calls its convolve method). The FFTConvolver (only
available if FFTW is found) uses an FFTPlan to carry out convolution. It can be
created either with an existing plan, or it can be instructed to create its own
plan, which can then be accessed.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>