Simplifying FFT reusability
What we want to reuse is not actually the plan, but the FFT convolver object,
which in turn uses a particular plan. Thus, the plan is now a unique, private
pointer held by the convolver, and we expose the convolver object via the model
instead for users to keep a hold of, and reuse (by giving it to a different
model of same characteristics).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>