Option to reuse FFT'd PSF
This saves as a forward FFT when doing more than one convolution using the same
PSF. The FFT'd PSF is still kept internally as part of the FFTConvolver, which
in turn is what can be shared across different Models. Although this design is
not as flexible as it would be to expose the FFT'd PSF via the convolver, users
probably want to reuse not only the FFT'd PSF but actually the whole convolver
anyway (i.e., its FFTPlan too).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>