Proper normalisation
The previous way of calling this meant that we were performing integer
arithmetic, and thus all values under 1 were rounded down to 0, resulting in a
sum of 0, and in no normalization!

This was never actually seen in reality by our ProFit users because they
normalize the PSF before it even gets to libprofit, so our
no-actually-normalisation normalisation didn't affect them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>