New _2ddata, Mask and Image classes
These classes encapsulate two-dimensional data and the different operations
that can be done with them. In particular, Image adds new numerical operations
that are not available in the base class (e.g., getTotal or normalize), and
that therefore Mask doesn't have.

The methods and operator overloadings of the Image class are meant to replace
both the code under utils.cpp (add_images, extend, crop, normalize) and other
places throughout the code where we used to perform image-related operations on
one-dimensional double vectors.

Tests are included for the new classes to ensure they work as expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>