Initial implementation of OpenCL 2D convolution
These kernels implement brute-force convolution for a single pixel of
the output image. The code follows the same logic as that on
convolve.cpp, where we implement convolution entirely on CPU. As usual,
we offer two flavours of this convolver (float and double).
The logic of the kernel is the same as that of the CPU brute-force
convolver, but applied to a single output pixel.

In the future I'd like to add a version that also uses local memory for
caching the values of the input image (and maybe also of the kernel).
This should yield better performance than what we currently are doing,
always fetching values from global memory.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>