Reworking convoultion for odd kernels
This was a long-standing issue with the code that never really saw the
light of day out in the open because ProFit (our main user) always
passes down an odd-sized PSF.

THe code was a bit weird because it was calculating half_krn_width as
(krn_width - 1 / 2), and the same for the height, and then using these
values to perform memory allocations, calculate offsets, etc. The code
yields correct results for odd-size kernels (because integer division
always rounds down), but yielded a not-so-correct result for even
dimensions (e.g., with PSF with = 10 it yields half_width = 4). The
final effect of this was that the convolution's result was offset by a
pixel on the even kernel dimension cases.

This patch fixed the calculation of this half_kernel_width/height to be
simply the dimension divided by 2. The FFT convolution needed to be
adjusted of course to yield the correct offsets as well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>