Getting gcc-4.6 from apt
Travis uses the Ubuntu Trusty images by default now, which don't include
gcc 4.6 anymore. We still want to make sure libprofit compiles there
though, and therefore keep that item on the matrix. To keep it we thus
need to pull gcc 4.6 from the repo (seems to be there) before using it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>