Redesigning convolver creation
The old way of creating convolvers was either by hand, calling the
corresponding constructor, or by creating a model, setting all the
appropriate fields on it, letting it create a convolver, and getting it.
This was far from ideal though, as it made it very difficult to know
exactly how to create convolvers.

A new centralized method of convolver creation is now available. Users
specify the type of convolver they want (either via string, or with an
enumeration), plus a set of creation preferences; then the appropriate
constructor is called and the convolver is returned.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>