Properly index devices while querying them
This was never an issue in the platforms that I had tested previously,
because they had only one device at most. When more than one device is
present in the same platform their information gets overwritten in the
map, leading users to incorrect information gathering, and device
selection later on.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>