Minor changes on how structs are initialized
Empty constructors now mean that members are not initialized. This means
that if we do a vector.resize() the memory will be allocated, but the
objects will not be initialized, which avoid touching those bits of
memory (and that we touch anyway later on, when filling the vector's
elements with actual values).

We are now also reusing the same ss_kinfo vector (resizing it) instead
of creating it every time on each loop iteration.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>