New brute-force convolver
This is the code that Dan wrote initially, but with more explicit
relaxed floating point operations associativity, and placed in a new
class of its own. At the very bottom of all the loops sits a dot product
operation. I couldn't think of a smart way to ensure that the data that
goes into the dot product is properly 16-bytes aligned (without copying
it over to an aligned buffer, that is), so the resulting machine code
does not fully use the XMM* and YMM* registers as it could if everything
was properly aligned.

Still, this code runs up to ~3x faster than the old BruteForceConvolver
code, which is quite a gain. A unit test ensures that the values
produced by this convolver are correct (which we deduce by comparing
them with what the old BruteForceConvolver produces).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>