Model::evaluate returns an Image
This is better than just returning the underlying data because it allows
us to return images with a size different from the requested one (which
could happen because a convolver doesn't crop an image after extending
it for convolution, for instance).

A type-cast operator in the Image class still allows us to directly
assign instances of Image to std::vector objects.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>