Improved image-related classes
First of all, a new _2dcoordinates class (typdef'd to Point and
Dimensions) eases the representation of 2D pairs. These two classes are
now used throughout the code (not in all places yet, but little by
little) to ease development and maintainance. A few unit tests check
that its functionality works as expected.

Secondly, a new surface_base class models the template-free aspects of
the old _2ddata class, which contains basically only the dimensions of
the surface. The _2ddata class has been now renamed to surface, and
accepts as a template parameter not only the datatype for the underlying
data, but also the derived type. This was needed to move the crop and
extend functionality to this base class, which returns the correct type
depending on whether it is invoked on a Mask or an Image, the two
subclasses of surface.

This new code therefore allows Masks to be extended and cropped, just
like Images can. The only extra functionality of the Image class now is
that it knows it can perform arithmetic operations over its values
(total, normalize, etc), which is not possible with Masks.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>