Convolvers can return uncropped result
A new parameter in the convolve method allow users to specify that the
internal result of the convolution implementation, which is usually
cropped and offset to be of the same size and on the same position of
the original image, should be returned without cropping. An extra
out parameter also allows users to find out, if the image is being
returned uncropped, what is the offset at which the cropping would have
occurred (therefore giving them a point of reference in the bigger
image).

The Model class is now using this during the evaluation process. The
crop flag is still hardcoded to true, but the rest of the logic is in
place to properly handle the case in which we don't want cropping, but
we still need to continue adding images on top of the convolution result
(because they didn't ask for convolution).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>