Model has option to return uncropped image
Because now the image returned by evaluate might be bigger, we also need
to return to the user the offset at which the actual requested image is.
We do this by changing the signature of the evaluate function, which now
retains an Image/Point pair.

profit-cli is now also aware of this new functionality, and it puts the
correct information in the fits files it outputs; this way fits viewers
should show the 0,0 coordinate where it corresponds. The unit tests have
also been ported to build and run correctly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>