Brute-force convolvers were wrongly using the mask
This was an error introduced during the porting to using the new Image
and Mask classes, but only got detected now by Dan when using ProFit
because we had no unit test to check this. The problem is now fixed, and
a new unit test ensures this doesn't occur again.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>