Very initial implementation of kernel cache
We are storing the binary version of the OpenCL program in different
files/directories for each platform/device/version combination, which we
load if found, instead of re-compiling the sources.

This is not taking into account yet that sources can change, and
therefore there is still working missing in that direction. In
particular, we need to checksum the sources, and compare that against
what's on disk to know if the specific cache entry needs to be
re-generated. If we want to be super nice, then we will do this in a
per-module basis, so only those modules whose source code has changed
need to be re-compiled and re-cached.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>