Use device's OpenCL version, not platform's
This was probably the cause of some problems I had in the past when running
tests in hyades. In that machine, the OpenCL platform version (CUDA) was 1.2,
but I never bothered checking the device's version (which as of last commit we
show with profit-cli). The code on the other hand was taking decisions based on
version, and was probably trying to execute OpenCL 1.2 commands on a device
with version < 1.2.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>