Initial appveyor support
We use appveyor to make sure libprofit builds correctly under Windows
(using the Visual Studio compiler). We use the vcpkg tool to install
both GSL and fftw3, which are required and optional, respectivelly. We
manually also clone the cxxtest sources, needed to build the tests. We
then use cmake to generate a VS solution file, run it to build libprofit
and its tests, and finally run the tests.

Most importantly, the generated profit.lib library needs to be in the
PATH for the tests to find it and run correctly.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>