Adjusting CmakeLists.txt for MSVC builds
Two minor changes are needed to properly support MSVC builds. First, we
are not building profit-cli at all, since it would require a bit more of
effort to get it into proper shape. Second, we pass down a specific
macro to the MSVC compiler to avoid getting too many warnings about some
of the standard library functions.

Other than that, no further changes are required and we obtain a proper
VS solution file out of cmake.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>