OpenCL header split in two
Now users of the library don't need to be aware of the internal details
of the OpenCL classes, which in turn means they don't need to pass any
specific flags to compile *their* software against an OpenCL-aware
libprofit installation.

The clear separation means that we offer a base OpenCLEnv class (that
now also allows them to query the device/platform name and version), but
we internally use much more functionality.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>