Redesigning public Model API
This is a big change. Its aim is to hide whatever happens with the Model's
attributes behind public API method (like it always should have been!).

Notably, the evaluate() method has also been re-designed back to its origings.
Most of the times users won't need to get an offset, so then making that offset
part of the return value is unnecessary. This brings us back to returning an
Image instead of a std::pair, which was cumbersome to deal with.

Since this change is quite disruptive, I also took the opportunity to move the
rest of the attributes to use the new types (Image, Point, etc) instead of
continuing using the old, more verbose types. This actually reduces the amount
of code that we need to write, specially in the tests.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>