Passing down PixelScale and magzero explicitly
This de-couples the Profiles from the Model even further (but not fully yet).
This paves the way to finally introduce finesampling properly into the Profile
classes, which will be easier once the Profiles are able to work on their own.

I also took the opportunity to add some missing documentation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>