Initial implementation of finesampling
Finesampling means that profiles receive a bigger image with a lower pixel
scale. On top of that, some profiles need to be informed about the finesampling
so they adjust their internal parameters. So far this is only the case for the
sky profile, but other might come.

A basic unit test is included, but more may come.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>