FFT header split in two
This follows the same logic used for splitting opencl.h and convolve.h; namely,
to isolate users from the internal workings of libprofit. With this separations
users don't need to include fftw.h anymore when building software against
libprofit.

An effect of this separation is that the effort_t enumertion doesn't live
inside the FFTPlan class anymore (because we actually want to hide the class
form the user). Thus, code using this enumeration needed amendment.

I also took the opportunity to clean up the documentation, making Doxygen
happier.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>