cxxtestgen needs care before cmake 3.7
The effects of this bug can be seen, for example, with brew-installed cxxtest
packages in MacOS, which places a cxxtestgen executable in the PATH that is not
a python script but a shell script. Other environments could see the same
problem though, so we address it generically (like it was addressed in upstream
cmake).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>