Deal with missing OpenCL version macros
Some environments (at least this has been observed in some old MacOSs) have
valid OpenCL headers and libraries, but no valid OpenCL version macro can be
found in the headers. In reality this means that we can't enable OpenCL.

This problem in turn flagged a problem in how we were enabling the
feature-specific tests in the test/ directory, which now use the appropriate
variables to check for feature support.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>