r2c and c2r transformations used for convolution
Our original implementation used the complex-to-complex FFT
transformations in both forward and backward direction to perform the
convolution of images. Because the images' data is actually made out of
real data (and not complex numbers) we can actually perform
real-to-complex and complex-to-real transformations instead, saving us
some memory and some CPU as well. Speedups are between 50% and 100%
compared with the previous implementation. All tests pass too.

The old complex-to-complex transformations are still being kept because
they will probably be useful to implement other logic. We still probably
want to separate them into a different class though, and rename them
something like FFTC2CPlan and FFTR2CPlan.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>