FFT classes redesigned
The previous FFT class, named FFTPlan, was becoming too heavy for
itself, encompasing too many use cases and using more resources than
needed (i.e., covering at the same time Complex-Complex-Complex
transformations and Real-Complex-Real transformations). Thus, the class
has been separated into two different classes to handle these cases
separately, and then use one or the other depending on the needs. So
far, the only user is the FFTConvolver which uses the R-C-R transformer,
and therefore the C-C-C transformer has been left commented out after
making sure it compiles (this is to not decrease coverage with code we
know is not used, but that will probably be used in the future).

The class names also do not have "Plan" in them anymore. This is because
even though they internally wrap one or more plans, their real objective
is to transform data.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>