Implementing FFTW wisdom import/export
FFTW wisdom is gathered automatically by FFTW when plans are created.
This wisdom can be exported and then imported, saving us time later when
creating new plans.

The wisdom is kept in different files depending on whether threading
support (i.e., OpenMP support) is available in the FFTW library we use
(in which case we always turn it on) or not. This is because "threaded"
wisdom is not loadable by an FFTW runtime that hasn't initialized its
thread support, and vice-versa.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>