Testing library-wide methods
We are now more interested the init and finish library-wide methods
because they actually have more logic inside them compared to before.

In reality we should be calling these during setUp and tearDown on each
of the tests, but we have been lazy enough to not do it, and probably
will stay like that for a while.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>