Improving documentation
Now that we have more methods at the library method, we need to properly
document how to use them and what they do. We are thus now making sure
their doxygen documentations are more useful, and including them in the
api.rst document, and also mentioning the init/finish pair in the
usage.rst document.

These are only docs, so let's [skip ci]

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>