Avoid warnings due to OpenCL > 2.0
This is happening in newer platforms, and causes lots of warnings due to
the cl2.hpp header saying that it doesn't support that version (which is
true), and stating that it will default to use 2.0 instead. This is fine
in terms of the result, but we can do without the warnings, so it's
better if we apply this limit ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>