Adding return_fine option to Model class
Setting this option to false (default is true) makes the Model return a
downsampled version of a finesampled image (i.e., it only makes sense
when using a finesampling factor > 1) such that the dimensions of the
image correspond to those originally set on the model. If there is any
offset caused by the convolver, the offset is also adjusted by the
finesmapling factor.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>