PROFIT_BUILD macros is not necessary anymore
This was previously used to identify when include files were being
included as part of the compilation of libprofit itself and when
external users were including them. The distinction was necessary
because we did not hide yet fully the implementation details from the
user-exposed header files, which we do now.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>