PROFIT_API-ing libprofit
This is required to properly support using the dllimport/dllexport
declspec specifiers, and thus correctly build a shared library under
Windows. We use the profit_EXPORTS macro already automatically defined
by cmake to distinguish between build-time include of headers (in which
case we declare dllexport) and user inclusion of headers (in which case
we declare dllimport). Our unit tests are already users who include the
headers with dllimport, so we make sure we ensure users can build
programs against our PROFIT_API'd libprofit already.

We still get many warnings due to std classes not being dllimported or
dllexported, but things seem to work despite this. I'm sure that in the
future if we want to get more serious we can deal with those warnings as
well.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>