Windows version of filesystem utils
This is not super advanced, but probably doesn't need to be. On the
other hand I didn't want to either pull boost-filesystem or c++17 just
to avoid writing this.

In the future if the code starts to get more bloated with #ifdefs we can
think on separating it into its own module.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>