Remove usage of rand_r
This is a POSIX function, and therefore not available in Windows. The
most optimal solution for fixing this problem would be to use the c++11
classes found in <random> to get random seeds, a number generator, and
then a uniform distribution. On the other hand this is not the only
place where we generate random data for our tests, so we'd probably want
to tackle all those places at once, and separately from the particular
problem this commit is fixing, so such transition is simply stated as
future work.