Moving data() members down to Image
This is because vector<bool> doesn't have a data() member, and therefore
trying to do Mask.data() fails. Our code does not do this directly, but
when trying to compile with MSVC (which installs a compiled version of
Mask into the dll) it fails -- and it would have failed otherwise if we
ever tried to do it ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>