Using hermitian size for complex vectors
During the real-to-complex and complex-to-real transforms, FFTW takes
advantage of the hermitian redundancy of the transformation, and
therefore needs only (N / 2 + 1) complex numbers to hold the
transformation of N real numbers. We were blind to this fact, and were
still handling complex vectors of size N. This meant that at convolution
time we were multiplying more complex numbers than necessary, and that
those numbers were uninitialized.

We now calculate the hermitian size that complex vectors should use, and
actually use it. Simple tests show a speedup of somewhere between 10%
and 20% in optimized builds depending on the CPU.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>