Reorganizing common code
Some of the protected methods in FFTTransform didn't really need to be
in the class definition. This is even more true now that they don't
actually use the private members anymore, so let us take them out of
there and into fft.cpp

Also, the way we were filling the vectors was ackward -- default
initializing their members doesn't really hurt, so let us do that
instead of creating/reserving/inserting.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>